Implement #20164: Hint and solution identifiers for questions with multiple hints
Changed the hint icon, that used to be a lightbulb, to display
the hint number directly. Providing a clear visual identification
of each hint number.
Each hint is prominently marked with its corresponding number
within the hint body, and it is possible to swap between hints' modals.
Changed the solution to be part of the "dialogue" between the
student and the chatbot.
All done following the proposed design by Oppia design team.

Co-authored-by: Ricardo Pereira  <ricardo.samora.pereira@tecnico.ulisboa.pt>